fix an issue when creating a new queteur,where an error messasge appeared, and the user needed to click one more time on save to actually save.
fix an issue on exportULData that fails because of $depart was added to the field list to export. Add a try / catch to avoid that this kind of error stops the export
ExportULData was exporting only one user, the nivol and the init_password_date was missing from data.

Fix various places to allow super admin to list troncQueteur of user outside its UL, list TroncQueteur History and see the detail of the tronc.

when publishing data to Google PubSub, put all the code that prepare the data in the try catch so that even if that fails, the end user doesn't get an error.

library update